fix(cloud-coding-agents): allowlist auth agents, log server errors, split settings UI
- server.ts auth route: validate agentType against a 4-item allowlist
  (claude/codex/gemini/aider) before passing to createAdapter. The
  \w+ regex on the URL blocks path traversal but still accepts
  arbitrary identifiers like __proto__, constructor, or any future
  adapter name the package happens to export. Gate on the shapes the
  UI actually ships today.
- server.ts auth route: log the full error (stack + message) via
  logger.error in the catch block before returning the generic
  client-facing message. The outer catch was correctly sanitizing
  the response but silently dropping production-debug info.
- Split CodingAgentSettingsSection.tsx (1064 LOC) into five smaller
  modules, bringing the parent to 444 LOC (under the ~500 guideline):
    - coding-agent-settings-shared.ts — types, constants, fallback
      model lists (128 LOC)
    - LlmProviderSection.tsx — LLM provider select + api_keys inputs
      + cloud pair status (174 LOC)
    - GlobalPrefsSection.tsx — selection strategy, approval preset,
      scratch retention, coding directory (199 LOC)
    - AgentTabsSection.tsx — segmented agent tabs, per-agent auth
      flow, auth-result banner, availability line (200 LOC)
    - ModelConfigSection.tsx — aider provider, powerful/fast model
      selects, fallback hint (150 LOC)
  Parent now owns state + load/save effects + handlers + render
  orchestration only. Update CodingAgentSettingsSection.test.tsx to
  import ADAPTER_NAME_TO_TAB from the new shared module.
